easy_xml/app/requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.1
djangorestframework==3.15.2
djangorestframework-xml==2.0.0
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.9
sqlparse==0.5.1
